requirements.txt
blinker==1.8.2
cachelib==0.13.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
Flask-Session==0.8.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3